refactor: move into module
