+++ equinox/_module/_module.py
@@ -1,7 +1,6 @@
 import dataclasses
 import functools as ft
 import inspect
-import textwrap
 import types
 import warnings
 import weakref
@@ -19,6 +18,7 @@ from .._pretty_print import tree_pformat
 from .._tree import tree_equal
 from ._better_abstract import better_dataclass, BetterABCMeta
 from ._field import field
+from ._flatten import generate_flatten_functions, WRAPPER_FIELD_NAMES
 
 
 # Legacy compatibility API, passed to `strict` below.
@@ -29,212 +29,9 @@ def StrictConfig(
     return None if force_abstact else False
 
 
-WRAPPER_FIELD_NAMES: Final = (
-    "__module__",
-    "__name__",
-    "__qualname__",
-    "__doc__",
-    "__annotations__",
-)
-
-
 _abstract_module_registry = weakref.WeakSet()
 _has_dataclass_init = weakref.WeakKeyDictionary()
 _module_info = weakref.WeakKeyDictionary()
-_MISSING = object()
-
-
-def _make_tuple_type(count: int, element_type: str = "Any") -> str:
-    """Generate a tuple type annotation string for a given count of elements."""
-    if count == 0:
-        return "tuple[()]"
-    elif count == 1:
-        return f"tuple[{element_type}]"
-    else:
-        return f"tuple[{', '.join([element_type] * count)}]"
-
-
-INDENT: Final = 4
-
-FIELDS_INFO = f"""
-    Dynamic fields: {{dynamic}}
-    Static fields: {{static}}
-    Wrapper fields: {WRAPPER_FIELD_NAMES}
-"""[1:-1]  # (trim leading and trailing newlines)
-
-
-FLATTEN_CODE_BASE = '''
-def flatten(obj: module_cls) -> {return_annotation}:
-    """Generated flatten function for {qualname}.
-    
-    {fields_info}
-    """
-    return {dynamic_vals}, {aux}
-'''
-
-
-FLATTEN_WITH_KEYS_CODE_BASE = '''
-def flatten_with_keys(obj: module_cls) -> {return_annotation}:
-    """Generated flatten_with_keys function for {qualname}.
-    
-    {fields_info}
-    """
-    return {key_tuple}, {aux}
-'''
-
-UNFLATTEN_FUNC_BASE = '''
-def unflatten(
-    module_cls: type[T],
-    aux_data: {aux_type},
-    children: {dynamic_type},
-) -> T:
-    """Generated unflatten function for {qualname}.
-
-    {fields_info}
-    """
-    self = object.__new__(module_cls)
-    {setters}
-    return self
-'''
-
-SET_DYNAMIC_BASE = """
-object.__setattr__(self, {name!r}, children[{i}])
-"""[1:-1]  # (trim leading and trailing newlines)
-
-SET_AUX_BASE = """
-if aux_data[{i}] is not _MISSING:
-    object.__setattr__(self, {name!r}, aux_data[{i}])
-"""[1:-1]  # (trim leading and trailing newlines)
-
-SET_WRAPPER_BASE = """
-if aux_data[{i}] is not _MISSING:
-    object.__setattr__(self, {name!r}, aux_data[{i}])
-"""[1:-1]
-
-SET_WRAPPER_LINES = "\n".join(
-    SET_WRAPPER_BASE.format(i=i, name=k) for i, k in enumerate(WRAPPER_FIELD_NAMES)
-)
-
-NS_BASE = {"object": object, "Any": Any, "tuple": tuple, "_MISSING": _MISSING}
-
-
-def _generate_flatten_functions(cls: type, fields: tuple[dataclasses.Field[Any], ...]):
-    """Generate optimized flatten/unflatten functions for a specific field config."""
-    # Separate dynamic and static fields
-    dynamic_fs_, static_fs_ = [], []
-    for f in fields:
-        if f.metadata.get("static", False):
-            static_fs_.append(f.name)
-        else:
-            dynamic_fs_.append(f.name)
-    dynamic_fs, static_fs = tuple(dynamic_fs_), tuple(static_fs_)
-    # aux_fs = WRAPPER_FIELD_NAMES + static_fs
-
-    # Build field info for docs
-    fields_info = FIELDS_INFO.format(dynamic=dynamic_fs, static=static_fs)[INDENT:]
-
-    # -------------------------------------------
-    # Generate flatten function
-
-    # Directly access dynamic fields by name
-    if dynamic_fs:
-        dynamic_exprs = [f"obj.{name}" for name in dynamic_fs]
-        dynamic_vals = f"({', '.join(dynamic_exprs)},)"
-    else:
-        dynamic_vals = "()"
-
-    # For static fields, we need to store their values in aux data
-    if static_fs:
-        static_exprs = [f"getattr(obj, {name!r}, _MISSING)" for name in static_fs]
-        static_aux = f"({', '.join(static_exprs)},)"
-    else:
-        static_aux = "()"
-
-    # Build return type annotation
-    dynamic_type = _make_tuple_type(len(dynamic_fs))
-    static_type = _make_tuple_type(len(static_fs))
-    qualname = cls.__qualname__
-
-    flatten_code = FLATTEN_CODE_BASE.format(
-        return_annotation=f"tuple[{dynamic_type}, {static_type}]",
-        qualname=qualname,
-        fields_info=fields_info,
-        dynamic_vals=dynamic_vals,
-        aux=static_aux,
-    )
-
-    # -------------------------------------------
-    # Generate flatten_with_keys function
-
-    key_exprs = [f"(jtu.GetAttrKey({name!r}), obj.{name})" for name in dynamic_fs]
-    key_tuple = f"({', '.join(key_exprs)},)" if dynamic_fs else "()"
-
-    keys_dynamic_type = _make_tuple_type(len(dynamic_fs), "tuple[Any, Any]")
-
-    flatten_with_keys_code = FLATTEN_WITH_KEYS_CODE_BASE.format(
-        return_annotation=f"tuple[{keys_dynamic_type}, {static_type}]",
-        qualname=qualname,
-        fields_info=fields_info,
-        key_tuple=key_tuple,
-        aux=static_aux,
-    )
-
-    # -------------------------------------------
-    # Generate unflatten function - directly set fields by index
-    # Extract types from flatten return type: tuple[dynamic_type, static_type]
-
-    unflatten_lines: list[str] = []
-    if dynamic_fs or static_fs:
-        unflatten_lines.append("# Set dynamic fields directly by index")
-    # Set dynamic fields directly by index
-    unflatten_lines.extend(
-        SET_DYNAMIC_BASE.format(i=i, name=k) for i, k in enumerate(dynamic_fs)
-    )
-    # Set wrapper fields from aux_data
-    # unflatten_lines.append(SET_WRAPPER_LINES)
-    # Set static fields from aux_data
-    unflatten_lines.extend(
-        SET_AUX_BASE.format(i=i, name=k)
-        # for i, k in enumerate(static_fs, start=len(WRAPPER_FIELD_NAMES))
-        for i, k in enumerate(static_fs, start=0)
-    )
-
-    unflatten_code = UNFLATTEN_FUNC_BASE.format(
-        aux_type=static_type,
-        dynamic_type=dynamic_type,
-        qualname=qualname,
-        fields_info=fields_info,
-        setters=textwrap.indent("\n".join(unflatten_lines), " " * INDENT)[INDENT:],
-    )
-
-    # -------------------------------------------
-    # Compile all functions
-
-    # Namespace for flatten functions (they need module_cls)
-    flatten_ns = NS_BASE | {"jtu": jtu, "module_cls": cls}
-
-    # Namespace for unflatten function (takes module_cls as parameter)
-    unflatten_ns = NS_BASE | {"type": type, "T": TypeVar("T")}
-
-    # Extract the generated functions from respective namespaces to
-    # set the proper module reference.
-    module_name = getattr(cls, "__module__", "equinox._module._module")
-    # flatten func
-    exec(compile(flatten_code, f"<generated_flatten_{qualname}>", "exec"), flatten_ns)
-    flatten_func = flatten_ns["flatten"]
-    flatten_func.__module__ = module_name
-    # flatten with keys func
-    func_name = f"<generated_flatten_with_keys_{qualname}>"
-    exec(compile(flatten_with_keys_code, func_name, "exec"), flatten_ns)
-    flatten_with_keys_func = flatten_ns["flatten_with_keys"]
-    flatten_with_keys_func.__module__ = module_name
-    # unflatten
-    func_name = f"<generated_unflatten_{qualname}>"
-    exec(compile(unflatten_code, func_name, "exec"), unflatten_ns)
-    unflatten_func = unflatten_ns["unflatten"]
-    unflatten_func.__module__ = module_name
-
-    return flatten_func, flatten_with_keys_func, unflatten_func
 
 
 MSG_METHOD_IN_INIT: Final = """Cannot assign methods in __init__.
@@ -426,7 +223,7 @@ class _ModuleMeta(BetterABCMeta):
         is_abstract: bool = False,
         strict: None | bool = False,
         **kwargs: object,
-    ) -> type["_ModuleMeta"]:
+    ):
         if strict is None:
             # Legacy compatibility API. Checking that this has the desired behaviour:
             #
@@ -515,7 +312,7 @@ class _ModuleMeta(BetterABCMeta):
 
         # Generate optimized flatten/unflatten functions
         flatten_func, flatten_with_keys_func, unflatten_func = (
-            _generate_flatten_functions(cls, fields)  # pyright: ignore[reportArgumentType]
+            generate_flatten_functions(cls, fields)  # pyright: ignore[reportArgumentType]
         )
 
         jtu.register_pytree_with_keys(
